+++ crates/meroctl/src/cli/context/common.rs
@@ -1,0 +1,21 @@
+use reqwest::Url;
+
+pub(crate) fn get_ip(multiaddr: &libp2p::Multiaddr) -> eyre::Result<reqwest::Url> {
+    let ip = multiaddr
+        .iter()
+        .find_map(|p| match p {
+            libp2p::multiaddr::Protocol::Ip4(ip) => Some(ip),
+            _ => None,
+        })
+        .ok_or_else(|| eyre::eyre!("No IP address found in Multiaddr"))?;
+
+    let port = multiaddr
+        .iter()
+        .find_map(|p| match p {
+            libp2p::multiaddr::Protocol::Tcp(port) => Some(port),
+            _ => None,
+        })
+        .ok_or_else(|| eyre::eyre!("No TCP port found in Multiaddr"))?;
+    let url = Url::parse(&format!("http://{}:{}", ip, port))?;
+    return Ok(url);
+}
